FARTEX-1: far textures above id 31128 were never released (16-bit truncation)
FarTexDB::Release bounded the texture id with a (WORD) cast:

  ShiAssert(texID < (WORD) texCount);

texCount is an int and is 96664 for the Korea theater, so the cast truncates the
bound to 96664 & 0xFFFF = 31128. TextureID is a DWORD and Request() already
compares against (DWORD) texCount, so ids above 31128 are loaded and refCounted
on request while Release() judged them out of range and returned early -- their
refCount incremented and never decremented. Those far textures were never freed.

Found by instrumenting the '[Failed: texID < (WORD) texCount]' assertion that has
been firing in every session rather than silencing it. The values gave it away at
once: every reported id was SMALLER than texCount, e.g. texID=33770 texCount=96664
'over by -62894'. A negative overrun is impossible unless the bound is not what it
looks like.

Very likely also the source of the long-recorded 'far texture loading errors
(42xxx IDs not found) - non-fatal': 42xxx sits squarely in the rejected range.

Verified on a 200s IA ASAN soak: assertions 2 -> 0, out-of-range rejections
10 -> 0, zero ASAN errors, 3D reached.

+++ src/graphics/texture/fartex.cpp
@@ -533,8 +533,18 @@ void FarTexDB::Release(TextureID texID)
 
     if (texID == INVALID_TEXID) return;
 
-    ShiAssert(texID >= (WORD) 0);
-    ShiAssert(texID < (WORD) texCount);
+    // FF_LINUX: (DWORD), not (WORD). texCount is an int and is 96664 for the Korea
+    // theater, so a (WORD) cast truncates the bound to 96664 & 0xFFFF = 31128 and
+    // every id above that is wrongly judged out of range. TextureID is a DWORD, so
+    // those ids are perfectly valid -- Request() below already compares against
+    // (DWORD) texCount and loads them. Only Release() was truncating, so far
+    // textures with id > 31128 had their refCount incremented on request and the
+    // decrement skipped: they were never freed. This is also the source of the
+    // "[Failed: texID < (WORD) texCount]" assertion in every session, and of the
+    // long-recorded "far texture loading errors (42xxx ids)" -- 42xxx is exactly the
+    // wrongly-rejected range.
+    ShiAssert(texID >= (DWORD) 0);
+    ShiAssert(texID < (DWORD) texCount);
 
 #ifdef FF_LINUX
     // FF_LINUX: see Request() - out-of-range ids corrupt texArray.
@@ -542,7 +552,7 @@ void FarTexDB::Release(TextureID texID)
     // fires in every session, and nobody has recorded HOW far out of range these ids
     // are or where they come from. That decides whether this is benign noise (like the
     // tex.cpp asserts) or a real id-space bug.
-    if (texID >= (WORD) texCount)
+    if (texID >= (DWORD) texCount)
     {
         static int ffRptd = 0;
